A-Packages: Restructure project into packages (commands, tasks, storage, ui, parser)
