+++ src/main/java/bestbot/command/DeleteCommand.java
@@ -1,4 +1,10 @@
-package bestbot;
+package bestbot.command;
+
+import bestbot.Ui;
+import bestbot.Storage;
+import bestbot.task.Task;
+import bestbot.task.TaskList;
+import bestbot.exception.BestbotException;
 
 /**
  * Command to delete a task by index.
