+++ src/main/java/bestbot/command/ExitCommand.java
@@ -1,4 +1,8 @@
-package bestbot;
+package bestbot.command;
+
+import bestbot.Ui;
+import bestbot.Storage;
+import bestbot.task.TaskList;
 
 /**
  * Command to exit the program.
